limiter + tests

+++ cmd/passer/log.go
@@ -29,7 +29,7 @@ func (run *Runner) logRuntimeInfo() {
 	run.logger.Info("Runtime Info",
 		zap.Int64("NumConcurrentRequests", run.concurrentRequests.Get()),
 		zap.Int("NumGoroutines", numGoroutines),
-		zap.Float64("MainLimiterTokens", run.mainLimiter.Tokens()),
+		zap.Int64("MainLimiterTokens", int64(run.mainLimiter.Tokens())),
 		zap.Int("NumCPU", numCPU),
 		zap.Int("GOMAXPROCS", gomaxprocs),
 		zap.Int64("NumCgoCalls", numCgoCalls),
